Add Workflow for Pull Request Analysis
The primary aim of this commit is to add the first GitHub Actions
Workflow to this project, which will run whenever a Pull Request is
opened in order to analysize the quality of contributions.

Additionally this commit also introduces the `main.py` file, and thus
begins to flesh out the project file strucutre. This also includes the
create of `environment.yml`, which is used to create virtual
environments via `mamba`.

Finally, some updates to the README in order to reflect issue story
point sccale.